Add picTypeForSharedLinking extension to CxxToolchain
Summary:
WTTS. Windows needs to force 'PDC' (position dependent code)
because there is no concept of 'pic' code on Windows. Technically, all
code for windows is supposed to be pic code, but because pic is as a
historical artifact the 'exception' (which is why we have static-pic as
a flavor), the default of pic causes breakages in windows abstractions
in buck.

The alternative would've been to change the one place where we throw
an exception in buck when pic mode is detected, but that actually
might cause more issues because we'd be allowing 'pic' builds on Windows
which are just a lie. So instead ensure we're preserving behavior.

fbshipit-source-id: 561a0359dbf906738544ae75987b901650c66a36