Add config to specify a target for proguard_config and optimized_proguard_config
Summary: We want to make it so that the entire Android toolchain is checked-in, and not dependent upon random paths on the host. This is a first step in that direction - we allow the user to set `proguard.proguard_config` and `proguard.optimized_proguard_config` which are specified as build targets with an output. The output is used instead of the random path on the host, if it is specified.

Reviewed By: sturmen

fbshipit-source-id: b86a0e77ac63ba098b93954c99d9b02779f02a19