add -fmodule-map-file when using explicit modules
Summary:
When compiling with explicit modules we need to pass up the module map files of the dependent modules being compiled so that Clang knows which module to load for a corresponding header. We can then disable implicit module map loading.

This fixes loading modules using the `-fmodule-file=name=path` syntax as well as avoiding issues with absolute paths being embedded in pcm files, as the command line overrides will take precedence over the serialized paths.

Reviewed By: milend

fbshipit-source-id: 38d82654b59df17a2e71ed6b0bee3fef25afc818