add test cases for platform_flagged_deps attr
Summary:
D34570245 is seeing weird errors and one theory was that this attr was breaking down. there were tests for the non-platform version of the flag but none for the platform one, so i added some to check if it was obviously broken (it's not)

we still don't know the cause of the problem in D34570245 but it appears to be a difference in how `CxxDeps` resolves its `PatternMatchedCollection`s vs how i did it in `RustLibraryDescription.java`. not sure we can trigger that edge case in this test harness

Reviewed By: jsgf

fbshipit-source-id: b8f58af58352e27da7449e93f44abb446d03533c