Get all manifests in the transitive deps
Summary:
There are two bugs here that are preventing us from getting all the manifests that we should be getting. The first is very old (introduced in D2414812) - we don't look at any of the transitive deps of `java_library`. Therefore if we have a graph like:
```
A(android_manifest) -> B(java_library) -> C(android_library with some AndroidManifest.xml)
```
then the AndroidManifest.xml is not included, when it should be.

The second bug is that `android_resource` does not return all of its deps, just ones that are on the Java classpath (which in particular does not include other `android_resource` deps). This was introduced by yours truly in D26436163.

This diff adds a config option to fix both of these bugs.

Motivation: `buck2` does this properly, and I want `buck1` and `buck2` to do the same thing, otherwise `buck2` runs into build issues (e.g. incompatible SDK versions) that `buck1` does not.

Reviewed By: jiawei-lyu

fbshipit-source-id: 509ac02c78d80eca696c2e26cf0b82c31d822412