use raw modules for cacheable pcm output
Summary: Use raw modules instead of obj modules. This will output a bitcode stream for the pcm file instead of an object file wrapper containing debug info. This seems to be the only way to prevent debug information being generated for pcm files, and the debug info generation does not respect path prefixing or the compilation directory. We should look at fixing this later.

Reviewed By: maxovtsin

fbshipit-source-id: c30e3221b0d56430697ce9a5d682934e558368b7