Make metadata storage configuration-safe
Summary:
Metadata storage has never been safe when a target is seen in
multiple configurations (even if that happens across different builds) because
the sqlite key is just the unconfigured target!

This has certainly been the source of a ton of unexplainable issues we've been
dealing with for years. For example:

missing/bad depfiles or other metadata
input-key mismatch
who knows what else

I actually wrote an integration test for this and verified it
manually running the buck commands from the integration test and then found
out that our integration test framework is also not configuration-safe (because
it depends on our build.log file which is not configuration safe)... so
that's all unfortunate.

Reviewed By: stepancheg

fbshipit-source-id: 57b2270dfeb87fa78c2e223011be15e0d137ca60