JavaCD refactoring
Summary:
Refactors javacd execution to do two main things:

1. execution context lifecycle management makes a lot more sense. It's really
strange that we were creating this in a function and then optionally either
returning it or closing it. Instead, it's now created, owned, closed all by the
caller and if the caller wants to keep it around and close it later its trivial
to do that.

2. returns the StepExecutionResult. previously, the execution stuff would
always write the result to the downward api pipes. this doesn't really work
outside the context of the worker tool. Instead, pull out that writing of the
result to the named pipe to the higher level.

Both of these are in preparation for a separate non-workertool entry point for
the javacd code.

Reviewed By: IanChilds

fbshipit-source-id: e3745ef9db597a0639106361db8b4b6b445d0645