Support buck1 bootstrap tools as prebuilts
Summary:
When iterating on the java rules, it can be really slow due to
invalidating the buck1 tools required for building and needing to then
rebuild those. This diff adds a configuration `-c
java.enable_bootstrap_prebuilts=true` that uses prebuilts in the source
tree and a script to build and update those prebuilts. This just
makes that development iteration a little better. Even better would
be if we could get the prebuilts and non-prebuilts case to have the outputs
look the same to dependents so that we can get cache hits with the non-prebuilt
case, but that's a bit trickier (though, i think, possible).

Reviewed By: IanChilds

fbshipit-source-id: cc2c0ca80aff7a672869603f0100f4ba0b375d3f