Allow relative paths in javacd.proto AbsPath
Summary:
buck2 avoids absolute paths a lot more than buck1 does and for buck2
integration the proto message is going to be an input to the action and so
really can't contain absolute paths. Instead, w/ buck2 absolute paths are
encoded as paths relative to the directory that actions are invoked in.

Add an env var that enables deserialization of that form of "absolute path".
This'll ensure that buck1 doesn't accidentally start using this path (which
in buck1 would not be cross-cell safe).

Reviewed By: IanChilds

fbshipit-source-id: 4f12039ef5a16c93f64874c1fab1787fee49e77c