buck: do not treat buck-out as symlink
Summary:
On windows whenever we try to serialize the action to send it to remote
execution we get the following failure:

{P500869937}

turns out this happens because we add files from buck-out directory, but also
we add buck-out itself as a symlink, because on windows with eden buck-out is a
symlink to a "scratch" folder. It seems like the easiest way to avoid this
problem is just not to treat buck-out as a symlink, and this is what this diff
does.

Reviewed By: KapJI

fbshipit-source-id: 892f49253d43a0a0fd151546712da09f08bceabb