added a temporary flag to enable explicit modules for clang
Summary: The diff adds a temporary flag in order to split clang explicit modules rollout from swift's explicit modules. The flag is added temporarily until clang explicit modules won't be tested on the same targets where they are enabled for Swift. Then the flag is going to be merged with uses_explicit_modules

Reviewed By: rmaz

fbshipit-source-id: d564f6777d34dc9645077291ea6c8a079f28e60a