Allow rustdoc_flags attribute for Rust targets
Summary:
I believe this needs to land (along with a buck1 version bump) before D36557638 would be allowed.

We're currently enforcing that buck2 and buck1 have the same attributes for all rules. That's done by generating https://www.internalfb.com/code/fbsource/fbcode/buck2/prelude/attributes.bzl from the set of attributes that buck1 claims it supports.

If `rustdoc_flags` is simply added to attributes.bzl without adding it to buck1 like in this diff, it'll be blown away whenever the file is regenerated, but until then, buck1 builds will fail with:

```
Error in rust_library: rustdoc_flags is not a recognized attribute
```

If `rustdoc_flags` is not added to attributes.bzl, then even if the support exists in buck2, buck2 builds in fbcode will fail with:

```
error: Found rustdoc_flags extra named parameter(s) for call to rust_library
   --> tools/build_defs/fb_native_wrapper.bzl:357:5
    |
357 |     native.rust_binary(**kwargs)
    |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
```

Reviewed By: zertosh

fbshipit-source-id: a823fb5c3c4faa06da7d7021497f897e01d47f22