Rename BaseJavacToJarStepFactory -> DaemonJavacToJarStepFactory
Summary:
`KotlincToJarStepFactory` currently creates a `JavacToJarStepFactory` within its
`createCompileSteps`, if it needs to add a `JavacStep` to the build steps.

With KotlinCD, this will not be possible, because `JavacToJarStepFactory` is the
factory that needs access to state that's held in the Buck process, while the
compile steps may be being created in the worker process.

Instead, we could create a `BaseJavacToJarStepFactory`, which is the factory
that JavaCD uses in its worker.  This doesn't need access to buck process
internals, but it does need to be supplied a `spoolMode`, which is used to
handle pipelining, but is not necessary for the purposes that the kotlinc
factory will use the javac factory for.

The solution is to create *another* class in the Javac Factory hierarchy, so it
becomes:

- `JavacToJarStepFactory`:  Aware of buck process internals, creates extra
  params.
- extends `DaemonJavacToJarStepFactory`:  Used by JavaCD, unaware of buck
  process internals, but aware of spooling.
- extends `BaseJavacToJarStepFactory`:  Only capable of creating compile steps
  (no pipelining, no buck process internals), shared between the javac and
  kotlinc step factories.
- extends `CompileToJarStepFactory<JavaExtraParams>`.

This diff renames the current `BaseJavacToJarStepFactory` to
`DaemonJavacToJarStepFactory` to make room for this new base.

Reviewed By: IanChilds

fbshipit-source-id: 12f67b38a9a6db584004933ebf0b600734bf5351