Integrate kosabi with buck rule
Summary: Kosabi is a source-only-abi optimisation for Kotlin incremental compilation. We need to integrate it with the build with the `SOURCE_ONLY_ABI_FLAVOR`. We are using plugin approach to update AndroidLibrary with enabling and config plugin for it. In the KotlinC step we can find the config and add as plugin options.

Reviewed By: artempyanykh

fbshipit-source-id: 4825f5cf010ead1bb9bf9f1c0a068d0bb3d47a57