Fix location of dependency symlink tree
Summary:
D34652155 added this symlink tree, but it used `BuildTargetPath.getGenPath` directly
instead of calling through MBR's OutputPathResolver, resulting in a symlink dir that is part of the
output directory, but isn't tracked as a produced artifact. We noticed this because outputSize was
logged inclusive of the symlink dir when the rule is BUILT_LOCALLY but does not include the symlink
dir when fetched from cache. This causes noise on the order of 400MB in the outputSize metric.

This diff fixes the issue by moving the symlink dir to the scratch dir (under `buck-out/bin`) and
uses the OutputPathResolver to do so.

Reviewed By: snufkinsnorka

fbshipit-source-id: 2f66d80b01fb24fe82b7cd05e9c6d37d8da060c7