Specify which files go into the secondary dexes too
Summary:
For Android APKs, we want to put some classes in the primary dex, and some into any one of the N secondary dexes that we produce.

When we have pre-dexed libraries, some of the those will have some classes that need to be in the primary dex, and some that go into the secondary dex. To tell `d8` which classes need to go into the primary dex, we pass it a list of the classes that go into the primary `classes.dex`, and then it puts the other classes into some secondary dexes.

My previous attempt to make this work assumed that the "leftover" classes (classes that are in the same jar as something in the primary dex, but that shouldn't be in the primary dex themselves) from the primary dex would only fill at most one secondary dex. However, as T121847925 shows, that assumption doesn't hold, and extending the same solution to multiple "leftover" secondary dexes is not easy (e.g. it's not obvious how to put a canary class in each).

Instead, for every DEX that we create (both primary and secondary), we specify which classes we want to be included, and ignore everything else. That way we should specify each class exactly once, and put it in exactly the DEX that we want it to belong in.

This simplifies `dex_rules.bzl` somewhat as we no longer have any special code for the primary dex "leftovers", just a primary dex and N secondary dexes.

Reviewed By: mykola-semko

fbshipit-source-id: 5a266210bafeac60b208d53ca8aec0224c312015