rename isSourceAbi -> isNotClassAbi
Summary:
Java so-abi optimisation happened after Java source-abi.
Most of the codebase assumes source-abi flavour is available if so-abi is available.

But It's not correct in Kotlin.
The following diffs detach `isSourceOnlyAbi` check from `isSourceAbi`

This diff renames `isSourceAbi` to what it actually mean

NOTE: it's not possible to build `stubs-gen` and `jvm-abi-gen` b/c action graph will have a cyclic dep (see `Kosabi.java`). We're going to fix it

Reviewed By: IanChilds

fbshipit-source-id: eddf29a27c5646633a7c6cb98c035a5a84fd2c19