Allow Overrides of JARs found in KOTLIN_HOME
Summary:
Add an ability to override JARs that are usually found in KOTLIN_HOME, via
buckconfig.  Following this diff, JARs are resolved as follows:

- If a source path is provided in buck config, this must exist, and this will
  be used (new as of this diff).
- Otherwise, look for the JAR within
  - `$KOTLIN_HOME`
  - `$KOTLIN_HOME/lib`
  - `$KOTLIN_HOME/libexec`

Where `$KOTLIN_HOME` is one of:
- `$(config kotlin.kotlin_home)` (which must be valid, if it exists).
- `$KOTLIN_HOME` (the environment variable), if it exists.
- The directory or parent directory of the `kotlinc` found in `$PATH`.

This diff does not get rid of the `kotlin.kotlin_home` config, to minimise
disruption across a buck bump, but I would also propose keeping the config, to
handle simpler integrations where people are happy to use a single `kotlin_home`
directory.

This diff also does not support using targets in these overrides, this is to
come in a later diff.

Reviewed By: mykola-semko, rybalkinsd

fbshipit-source-id: b5083f2917243907add3a67efb00b86aefe45dbf