thin archive step should run locally, fix conditional remote execution for CxxLink
Summary:
buck1 materializes all outputs locally anyway, so it doesn't make sense to
run a thin archive step remotely. This is especially noticable and impactful for Windows
RE where materializing inputs is particularly expensive and would grossly dominate the
execution time of the archive step (which, for thin archives, is generally under a second).

Also fix the conditional remote execution for CxxLink which wasn't on the Buildable like
it should be while I'm at it.

Reviewed By: mykola-semko

fbshipit-source-id: 82c881b868b3130e921ff222a9eb1543f7c4da52