populate SONAME field in shared library for all Rust shared libraries
Summary:
i don't understand why this was limited to `cdylib` before, if it's important please enlighten me. some discussion [here](https://fb.workplace.com/groups/llvm.gcc/posts/8638060492909156) - we think it's the right change

D36335572 fixes the problem of Rust `.so`s not being included in Android APKs, but they're still not loadable because the runtime is looking for them in a `buck-out` path. but if we set the `SONAME` field in the Dynamic Section then it loads them from the right place

D13055921 populated SONAME in the first place and introduced this cdylib limitation. no mention of why, probably just trying to minimally change behavior to fix the bug

Reviewed By: jsgf

fbshipit-source-id: 750fb6aa77430d268ebd3952cd40f3d3a9b0816c