buck1: report accurate cache sizes for two-level content lookups
Summary:
When we do a two-level content lookup we report the size of the "first" entry
we look up, but this is always zero, so it's not at all helpful or accurate,
this means that our Scuba metrics and our console end up showing inaccurately
low numbers (like, by about an order of magnitude).

Looking at places that use this in the codebase it doesn't look like this is
used for anything but logging so I guess this is safe?

Reviewed By: ndmitchell, StanislavGlebik

fbshipit-source-id: 89146e8b377ffca84cca3146e3a1dfeed2c5144f