report separate cas download metrics for artifacts cache and RE
Summary:
We are starting to use CAS as buckcache artifacts cache. We want to separate
CAS download/upload metrics for RE and artifacts cache, because it makes
accounting much harder (for example, at the moment we can't get the total
number of downloaded bytes).

This diff adds a separate logging for buckcache and artifacts cache.

Reviewed By: miaoyipu

fbshipit-source-id: 79567aa7cb5b878035a492276fd08cae9b90a1a6