Push RulesCDParams creation down into ConfiguredCompilerFactory
Summary:
...after pulling it out of `DefaultJavaLibraryRules`.  This allows the choice
of compiler daemon params (which depends on which language is being compiled
and whether or not it supports a compiler daemon) to be restricted to within
the factory that chooses compilers based on language.

The params are also made accessible from `JarBuildStepsFactory`, which goes
wherever those params are needed, which helps avoid an extra parameter in a
number of places.

There are some hold-outs that need to create a `RulesCDParams` from scratch (to
do with AARs and BuildConfig.java) but they are already special cases.

This diff also cleans up some dead references to `BuckJavaCDConfig` within the
non-Java JVM language support, which was previously hidden.

Reviewed By: mykola-semko, rybalkinsd

fbshipit-source-id: 3e46571c23e25329ce050b681767606f769c2729