Remove references to BuckJavaCDConfig from DefaultJavaLibraryRules
Summary:
`BuckJavaCDConfig` is currently used by `DefaultJavaLibraryRules` to create the
compiler daemon params.  However, the former config is only relevant for Java,
and the latter class is used to build all different kinds of JVM libraries.

This diff pulls the creation of `RulesCDParams` out of
`DefaultJavaLibraryRules` so that the decision of which params to pass is made
at the same time as which language(s) are being compiled in this library.

Some (most) languages do not support compiler daemons, so a sentinal
`RulesCDParams` is also introduced with all disabled/false flags set, to be
passed in for those languages.

A further improvement, in an upcoming diff, is to make the creation of
`RulesCDParams` the responsibility of the configured compiler factory (which
is also tied to the language being compiled).

Reviewed By: IanChilds, mykola-semko

fbshipit-source-id: e9db794b1337eb3a5a572a07fdcfbe76f2283256