Bring creation of Java-specific CD Params into java package
Summary:
As in title.

This utility method (`JavaCDParams.get`) is used in the implementation of
`DefaultJavaLibraryRules` which is used for other languages.  In a future diff,
I will remove the Java-specific references from there, to make it possible to
pass different cd params in based on the language that is being compiled.

Reviewed By: IanChilds, mykola-semko

fbshipit-source-id: 4c0a4f9176882a38aebb1221b3008143228a1bd8