Migrate project components to project service
Summary:
According to this page https://plugins.jetbrains.com/docs/intellij/plugin-components.html, the components are deprecated and we have to switch to project service or listeners etc.

This diff migrates below project components to project service:
* BuckCellSettingsProvider
* BuckExecutableSettingsProvider
* BuckProjectSettingsProvider
* BuckCellManager
* BuckTargetLocatorImpl

The BuckModule is a more complicated one so will handle it in next diff.

Reviewed By: cute-jumper

fbshipit-source-id: df9f52e8cbd86aae1f2b241bfbd685cd8716081e