Don't throw if unaccounted for depfile entries are considered uncovered.
Summary:
If we have, say, a genrule which emits/contains a sysroot of headers and
tools, and we pass `$(location //:sysroot)` in the args of a command alias that is
used as, say, a cuda preprocessor, this path gets added to a 'coveredPathsPredicate'
in the depfile logic. This causes an early exit from depfile calculations. However,
if we also pass the same `$(location //:sysroot)` through a `-I` flag in, say,
cudapp_flags, this path ends up in the depfile inputs but does not get properly
accounted for. Post compilation, the depfile logic will throw an exception if all
depfile paths were not accounted for. We should skip those paths which were handled
elsehow.

Reviewed By: mzlee

fbshipit-source-id: 4d67380d49c1c31d6e1ca269c42d0aea2ab5679b