buck install: Extend install.proto with ErrorDetails
Summary:
^
In case of an error we need to populate error detail field, otherwise leave it empty (null). No need to keep 2 other fields :  `err` and `err_msg`.
Also in case we want to extend error details with extra fields (like map of metadata associated with error) we could easily do that as we are now having a separate `message` for it.

In the next diffs I will switch to this new field and remove existing `err` and `err_msg`

Reviewed By: lebentle

fbshipit-source-id: 3fee593e2f720e6eedd3dc5419b50b3bd9fa0c0d