Choose kind of CompileStepsBuilderFactory to with dynamic dispatch
Summary:
`CompileStepsBuilderFactory` is the interface that allows us to switch out
steps to run in the main buck process with steps to run over JavaCD.
Previously, it was created by `JavaCompileStepsBuilderFactoryCreator` which
decided which kind to produce by checking the type of the configured
`CompileToJarStepFactory` (if it was the one that supported JavaCD, then use
that builder, otherwise, use the default).

This was difficult to extend to Kotlin, so this change introduces an API to
`CompileToJarStepFactory` to decide the kind of `CompileStepsBuilderFactory` to
use when interacting with it.

The rest of the changes in this diff are to do with moving classes between
packages and buck targets to avoid buck dependency graph cycles:

- Splitting `CompileToJarStepFactory` into the interface that
  `CompileStepsBuilder`s interact with, and the existing base implementation
  (now called `BaseCompileToJarStepFactory`).
- Move the new `CompileToJarStepFactory` interface to `.../jvm/java:support`
  (previously `.../jvm/java:steps`) to avoid a cycle between `.../jvm/cd` and
  `.../jvm/java:steps`.
- Moving the `CompileStepsBuilder` (+ `*Factory`) into `buck/jvm/cd` to avoid a cycle
  between `.../jvm/cd:cd` and `.../jvm/cd:steps`.

Reviewed By: mykola-semko

fbshipit-source-id: fc871e019955abfb005a1c2fd2920703ec4bf2cb