buck install: add Install rpc method to install.proto
Summary:
A new method that reflects install info provider from buck2.

The usage of this new method would be the following:
* buck2 calls `install` to notify installer app about requested install artifacts (install_id and map of files to install)
* When file is ready -> buck calls `file_ready_request` method
* At the end buck calls `shutdown_server` method

This change in the workflow will allow to support multiple install requests with the same installer app. For example `buck2 install m4a fb4a` could be supported with the same android installer app. Files that belong to m4a would have an installer_id associated with it, and files for fb4a would have a separate correspondent installer_id.

Reviewed By: lebentle

fbshipit-source-id: 4640e7ed37ceb916059362e6f21b1e8fd5279463