Add absolute path of project root to test running options
Summary:
Tpx/testpilot needs to report where tests live, for buck based tests, that will be the project relative folder path of the build (BUCK, TARGETS) file. We are making buck report this path to external test runners to avoid hacks on test runner side.

This is already done for buck2 in D37857557, so I'm just replicating that to buck1.

This diff just adds the absolute path of project root to test running options so that the project relative path of packages can be resolved when external test spec is generated

Reviewed By: bobyangyf

fbshipit-source-id: 0f9f5369e6687faee810d4994bf7b1fcc6cd122d