Allow packaging wrap shell scripts in APK lib directories
Summary:
Wrap shell scripts (https://developer.android.com/ndk/guides/wrap-script)
enable passing extra arguments when launching an Android app, which is
useful for e.g. getting ASAN libraries set up. We want to package them
in the app's native library directory alongside other native libraries,
so allow wrap.sh files to be placed in the native library directory.

The main motivation for treating the wrap shell script as a native
library is to get ABI filtering for it (i.e. respecting the cpu_filters
of the Android app). This is important because Android prefers 64-bit
libraries to 32-bit libraries. Our current setup involves placing the
wrap shell script in a prebuilt jar, but then we miss out on ABI
filtering, and that in turn prevents us from turning on ASAN for 64-bit
apps because of the 64-bit library directories getting included in
32-bit apps and causing crashes: D35723871.

Reviewed By: IanChilds

fbshipit-source-id: eb48480cd9aa54df072ad5ee1edfff5a9c477015