Make is_asset and has_wrap_script mutually exclusive
Summary:
It doesn't make sense to use both of them together; error out if that's
attempted.

The test added didn't quite fit into any of the existing integration
tests, and I'm testing error states so I was also hesitatant to
incorporate it into an existing suite for fear of breaking something.
prebuilt_native_library also lacks its own integration test, and it felt
strange to add one just for this one error state. I'll be adding
`used_by_wrap_script` to cxx_library as well in a follow-up, so I
figured I could create a new integration test for the possible errors
related to this argument for both rules.

Reviewed By: IanChilds

fbshipit-source-id: cf949fc0d1070465ad5454143c6f748bcc32ff43