Exclude system loader native libraries from instrumentation APK
Summary:
Match the logic for all other prebuilt native libraries. I wasn't sure
about the virtues of adding a new collection for the system loader
libraries to exclude versus just extending one of the existing
collections, but we have the precedent of the asset and non-asset
libraries having separate collections, so I added a new collection to
follow suit.

Reviewed By: IanChilds

fbshipit-source-id: 6e6972d1ff0b30dc063162dab713be8dac9c8ef6