fix issue when provided test external runner is exactly the same of the one provided via .buckconfig, and target not in prefix values
Summary:
When the test target isnt part of the prefix values, and the test.external_runner is matching the same one that was provided by the .buckconfig, the external runner would be ignored and the internal runner incorrectly used instead.
Because buck merges all the configs into a single one, and it doesn't store the overrides config in the config map, buck didn't have enough information to know if the user provided an external runner, landing on that invalid state.
This change is adding a new `cliOverridesConfig` parameter to the Config, storing the overrides from the user.  With that information, the `ExternalTestRunnerProvider` can now detect if the user has overwritten the external runner, and in the case of it being the same one that was defined by the combined buckConfigs, we return that defined runner as expected

Reviewed By: bobyangyf

fbshipit-source-id: 1a1065822a918dac941f8eceef70b2bb8c3293ee