Make APKModuleGraph more generic
Summary: Right now, `APKModuleGraph` is specific to `TargetGraph`, `TargetNode` and `BuildTarget`. However, we want to use it on "external" graphs (in particular we want to use it in `buck2`) and to do that we make it generic - it uses `TargetGraphInterface`, `HasBuildTargetAndBuildDeps`, and a parameterized type `BuildTargetType`. This means that we can build our own graphs and pass them to `APKModuleGraph` to be processed.

fbshipit-source-id: cf698999688bb2cf438f229559819a6e0f288515