Support xz compression of native libraries packaged as assets
Summary:
If `compress_asset_libraries` is `True`, then we collect up all the native libraries that are packaged as assets and compress them into a single file in `assets/lib`.

This adds that support for `buck2`.

Note that we need to know the exact path of each native library. Rather than search the filesystem, we already can easily get that information from the action that creates the metadata, and so we additionally output the exact paths from that action and use that.

#forcetdhashing

Reviewed By: smeenai

fbshipit-source-id: 49d3a65f95fb39f726a10f4956c87b013ae80ece