Split KotlincToJarStepFactory into regular and Daemon variants.
Summary:
The regular variant has the same behaviour as before.  The `Daemon` variant is
the one that is supposed to be used within KotlinCD.

This change exposes a couple more parameters that are being passed into the
constructor of `DaemonKotlincToJarStepFactory` which should be passed via
`KotlinExtraParams` instead.  This will be addressed in a follow-up diff.

Reviewed By: acejingbo

fbshipit-source-id: b0e3c521be11d2207043b3a06eb4613c2e26b4ab