fix testng requiring annotations to set alwaysRun=true
Summary:
TestNG tests using Before/After annotations requires to be set alwaysRun=true, because a Filter disable the annotations preventing it to be executed when TestRunner is executing with --filter
It's an issue that projects on fbcode live with, but onboarding projects that use testng from maven/gradle will break.

This fix will allow Presto to start using buck.

Reviewed By: ericyuliu

fbshipit-source-id: b3b7a562464c4998422da14e55090d1d5b831a30