fix testng requiring annotations to set alwaysRun=true
Summary:
TestNG tests using Before/After annotations requires to be set alwaysRun=true, because a Filter disable the annotations preventing it to be executed when TestRunner is executing with --filter
It's an issue that projects on fbcode live with, but onboarding projects that use testng from maven/gradle will break.

This fix will allow Presto to start using buck.

Reviewed By: ericyuliu

fbshipit-source-id: fe8c0c7e64e336358cdfbd428da074e424f711a7