implement AndroidPackageable for RustLibrary (reland)
Summary:
Original commit changeset: 03e7a83927cc

Original Phabricator Diff: D37194291

D36335572 implemented this, D37194291 reverted it because of build failures that arose with hypernova stuff

i think we've figured out the issues are solved by D38995945 and hypernova needs this diff relanded for merge maps support (and i want to land it to land my android contbuilds). so, relanding

also incorporates D37184285 which is a fix for a bug that was in the original diff.

Reviewed By: capickett

fbshipit-source-id: 6a5a2a95f2b36be30417936d861e7865559b99e9