changing voltron dex name to follow classesN.dex format
Summary:
were looking to change the naming convention of voltron modules to follow google standards. this means changing format moduleN.dex -> classesN.dex. side note, well need to follow up and move these dex files into a dex directory and make a change to the android manifest. but those changes are situational on a few configurations not passed to buck so we will modify in our own postprocessor. it seemed like we should rename in buck though since this should apply for all our modules.

the motivation behind this is that android OS can optimize and generate odex/vdex files for our ART builds such that our load times are improved.

this should not effect how dextricks and base class loaders to work

(Note: this ignores all push blocking failures!)

Reviewed By: IanChilds

fbshipit-source-id: 2c0f05c765aa9ea047229f2624270a61f644e947