Fix CxxInferCaptureRule invalidation on InferPlatform changes
Summary:
Apparently, because of `SupportsDependencyFileRuleKey` marker interface on
CxxInferCaptureRule and the surrounding logic, InferPlatform updates didn't
cause this file-capture rules rule-key invalidation.

(Looking at the depfile, there's a bunch of headers from infer's distro but no
binary or anything; so perhaps this is why the rule invalidation was borked).

I'm not sure if using `SupportsDependencyFileRuleKey` was intentional or just a
carry-over when the rule was migrated to MBR. But removing it fixes the problem
with bad cache/corruption on infer updates.

Reviewed By: ngorogiannis

fbshipit-source-id: 3cc97400f5a9aebe9f36ee7fad37566fa9a950d2