Add allFilesReady to InstallCommand API
Summary:
When we are doing an Android Install, we don't actually know exactly what files we're going to be given (we get different files depending upon whether it is an exopackage install or not).

Add an extra method that is called when all the files are ready - this means that we can simply do the install at that point. This also simplifies the `AndroidInstallerManager` code as it no longer has to track all of the files that it has received.

(Note: this ignores all push blocking failures!)

Reviewed By: lebentle

fbshipit-source-id: 390d28005817e05bbb652f3a245fc3042c39208c