Allow specifying that a build is for OSS via reading a buckconfig.
Summary:
We are trying to move tests to use Tpx for SandcastleBuckVerification. E.g jobs:
 - Linux: buck-verification-linux-java11
 - Linux OSS: buck-verification-linux-java11_opensource

Note that we have no way to differentiate from the build system if a release is for OSS or not. However, the CI jobs have that context.

To be able to propagate that to the test runner, we can have an extra config `-c build.is_oss=true` that will allow us to propagate an extra label.

Reviewed By: krallin

fbshipit-source-id: b25d14a45a2bcccf38d00f0cd5f3426509596a81