Fix RobolectricRuntimeDependencies rule
Summary:
This rule is currently unused, but we want to start using it to avoid over-fetching large android-all jars when running Robolectric tests.

I found that when buck found a cache hit for this rule, it did not produce the output symlinks, but these are always required at runtime by Robolectric. I think that making this rule uncacheable is the fix, and I was inspired by a rule which similarly creates symlinks:

https://www.internalfb.com/code/fbsource/[f227b3f45d842d7b01c9532579ea1f13890ec133]/xplat/build_infra/buck_client/src/com/facebook/buck/features/haskell/HaskellGhciRule.java?lines=577-583

Additionally, I found that when this rule is executed in RE the output symlink files do not appear locally. Seems like we just need to disable RE for this rule.

Reviewed By: IanChilds

fbshipit-source-id: c0d20a267d253093909f8fc72b00c9d201f8db19