buck1: quote classpath argsfiles
Summary:
Like it says in the title. If flavors are in used the classpath might contain a
`#`, which is a comment. It seems we can't actually escape those, but we can
quote them, so since we know the classpath is always supposed to be just one
argument, do that unconditionally.

Before:

```
Error: -cp requires class path specification
```

After: it works.

The docs do indicate that `"` is for quoting though it doesn't explicitly say
that comments can be escaped that way ... See the section called "java
Command-Line Argument Files"

https://docs.oracle.com/en/java/javase/15/docs/specs/man/java.html

Reviewed By: IanChilds

fbshipit-source-id: 059110aa08c03648ec2603b4ecac6d23917aad0e