Update ZipScrubber to handle files larger than 2 GiB
Summary:
ZipScrubber could not handle files larger than two GiB due to an underlying limitation in FileChannel. This adds a SlidingFileWindow that provides a ByteBuffer-like implementation. Calls are mapped to an underlying ByteBuffer that can be re-mapped to different parts of the file as needed.

Other changes were necessary to handle some of the facets of larger zip files, like using the ZIP64 headers and extras to identify the location of file headers.

Reviewed By: christolliday

fbshipit-source-id: 928f15da4551015fb68b1cc52ffd0f5be3af224a