Add support for enabling emitting textual objc headers on Swift toolchains that support it
Summary:
The pika toolchain has been patched to support a `-Xfrontend -emit-objc-header-textually` option
which tells the Swift compiler to generate ObjC headers fit for consumption without -fmodules enabled.
This change modifies the buck client such that if the active Swift toolchain supports it, the in-compiler method of generating textual headers is used over the buck post processing tool.

Reviewed By: rmaz

fbshipit-source-id: 9e28b087e57f1afa51596b48703cca6bbe2f8c98